Refactor health into separate file and fix aplite
