+++ src/windows/main_window.h
@@ -3,6 +3,7 @@
 #include <pebble.h>
 
 #include "modules/data.h"
+#include "modules/health.h"
 
 void main_window_push();
